+++ hopper_can_interface/CMakeLists.txt
@@ -1,26 +1,26 @@
-cmake_minimum_required(VERSION 3.0.2)
+cmake_minimum_required(VERSION 3.1.0)
 project(hopper_can_interface)
 
-set(CMAKE_EXPORT_COMPILE_COMMANDS TRUE)
 set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wfatal-errors")
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wfatal-errors")
 
-set(CMAKE_THREAD_PREFER_PTHREAD TRUE)
-set(THREADS_PREFER_PTHREAD_FLAG TRUE)
+# Pthread
+set(CMAKE_THREAD_PREFER_PTHREAD ON)
+set(THREADS_PREFER_PTHREAD_FLAG ON)
 find_package(Threads REQUIRED)
 
 find_package(catkin REQUIRED)
 
 catkin_package(
-  INCLUDE_DIRS 
-    include
-  LIBRARIES 
-    ${PROJECT_NAME}
+  INCLUDE_DIRS
+  include
+  LIBRARIES
+  ${PROJECT_NAME}
 )
 
-###########
-## Build ##
-###########
+# ##########
+# # Build ##
+# ##########
 include_directories(
   include
   ${catkin_INCLUDE_DIRS}
@@ -29,31 +29,30 @@ include_directories(
 add_library(${PROJECT_NAME}
   src/BufferedCanMsg.cpp
   src/CanInterface.cpp
+  src/ODriveCan.cpp
 )
 target_link_libraries(${PROJECT_NAME} pcanbasic Threads::Threads)
 
-add_executable(TestCanWrite
-  test/TestCanWrite.cpp
-)
-target_link_libraries(TestCanWrite ${PROJECT_NAME})
+# ############
+# # Install ##
+# ############
 
-add_executable(TestCanRead
-  test/TestCanRead.cpp
-)
-target_link_libraries(TestCanRead ${PROJECT_NAME})
+# ############
+# # Testing ##
+# ############
+if(${CATKIN_ENABLE_TESTING})
+  add_executable(TestCanWrite
+    test/TestCanWrite.cpp
+  )
+  target_link_libraries(TestCanWrite ${PROJECT_NAME})
 
-#############
-## Install ##
-#############
+  add_executable(TestCanRead
+    test/TestCanRead.cpp
+  )
+  target_link_libraries(TestCanRead ${PROJECT_NAME})
 
-
-#############
-## Testing ##
-#############
-
-# Add gtest based cpp test target and link libraries
-catkin_add_gtest(${PROJECT_NAME}-test 
-  test/test_hopper_can_interface.cpp
-)
-target_link_libraries(${PROJECT_NAME}-test ${PROJECT_NAME} gtest_main)
-
+  add_executable(TestODrive
+    test/TestODrive.cpp
+  )
+  target_link_libraries(TestODrive ${PROJECT_NAME})
+endif()
